refactor: AggregationContext→Aggregator、aggregator→aggregateにリネーム
クラス名をAggregatorに、型定義モジュールをaggregate.tsにリネームし、
ファイル名と役割の対応を明確化。全importパスを更新。
