+++ src/components/CrossConfig.ts
@@ -1,4 +1,4 @@
-import { questionKey, type QuestionDef } from "../lib/aggregator";
+import { questionKey, type QuestionDef } from "../lib/aggregate";
 
 export interface CrossConfigState {
   questions: QuestionDef[];
