+++ src/lib/duckdbBridge.ts
@@ -3,7 +3,7 @@ import duckdb_wasm from "@duckdb/duckdb-wasm/dist/duckdb-mvp.wasm?url";
 import mvp_worker from "@duckdb/duckdb-wasm/dist/duckdb-browser-mvp.worker.js?url";
 import duckdb_wasm_eh from "@duckdb/duckdb-wasm/dist/duckdb-eh.wasm?url";
 import eh_worker from "@duckdb/duckdb-wasm/dist/duckdb-browser-eh.worker.js?url";
-import { aggregate, type Query, type AggResult } from "./aggregator";
+import { aggregate, type Query, type AggResult } from "./aggregate";
 
 type DuckStatus = "loading" | "ready" | "error";
